# api/internal/handlers/default.go
package handlers

import (
	"api/internal/models"
	"api/pkg/services/worker"
	"log/slog"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/pgvector/pgvector-go"
)

const (
	minSearchQueryLen  = 3
	searchDefaultLimit = 50
)

// GetVersion ...
func (h *Handler) GetVersion(ctx echo.Context) error {
	version := models.GetVersion()
	return ctx.JSON(http.StatusOK, version)
}

// GetFeatures ...
//
//nolint:cyclop
func (h *Handler) GetFeatures(ctx echo.Context) error {
	cfgFeatures := models.GetFeatures(h.Config)
	features, _ := ctx.Get("features").(models.Features)

	features.Favourites = features.Favourites && cfgFeatures.Favourites
	features.Hidden = features.Hidden && cfgFeatures.Hidden
	features.Trash = features.Trash && cfgFeatures.Trash
	features.Albums = features.Albums && cfgFeatures.Albums
	features.Explore = features.Explore && cfgFeatures.Explore
	features.Places = features.Places && cfgFeatures.Places
	features.Things = features.Things && cfgFeatures.Things
	features.People = features.People && cfgFeatures.People
	features.Sharing = features.Sharing && cfgFeatures.Sharing
	features.Jobs = features.Jobs && cfgFeatures.Jobs

	return ctx.JSON(http.StatusOK, features)
}

// GetDisk ...
func (h *Handler) GetDisk(ctx echo.Context) error {
	disk := models.GetDisk(h.Config)
	return ctx.JSON(http.StatusOK, disk)
}

// Search ...
func (h *Handler) Search(ctx echo.Context) error {
	searchQuery := ctx.QueryParam("q")
	if len(searchQuery) < minSearchQueryLen {
		return echo.NewHTTPError(http.StatusBadRequest, "invalid search query")
	}
	mediaItems := []models.MediaItem{}
	if h.Config.Search {
		searchEmbedding, err := h.Worker.GenerateEmbedding(ctx.Request().Context(), &worker.GenerateEmbeddingRequest{Text: searchQuery})
		if err != nil {
			slog.Error("error getting search query embedding", "error", err)
			return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
		}
		rows, err := h.DB.Query(ctx.Request().Context(), "SELECT * FROM mediaitems WHERE id IN (SELECT id from mediaitem_embeddings ORDER BY embedding <-> $1) LIMIT $2", pgvector.NewVector(searchEmbedding.Embedding), searchDefaultLimit)
		if err != nil {
			slog.Error("error searching mediaitems", "error", err)
			return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
		}
		defer rows.Close()
		for rows.Next() {
			mediaItem := models.MediaItem{}
			if err := rows.Scan(&mediaItem.ID,
				&mediaItem.UserID,
				&mediaItem.Filename,
				&mediaItem.Hash,
				&mediaItem.Description,
				&mediaItem.MimeType,
				&mediaItem.SourceURL,
				&mediaItem.PreviewURL,
				&mediaItem.ThumbnailURL,
				&mediaItem.Placeholder,
				&mediaItem.IsFavourite,
				&mediaItem.IsHidden,
				&mediaItem.IsDeleted,
				&mediaItem.Status,
				&mediaItem.MediaItemType,
				&mediaItem.MediaItemCategory,
				&mediaItem.Width,
				&mediaItem.Height,
				&mediaItem.CreationTime,
				&mediaItem.CameraMake,
				&mediaItem.CameraModel,
				&mediaItem.FocalLength,
				&mediaItem.ApertureFnumber,
				&mediaItem.IsoEquivalent,
				&mediaItem.ExposureTime,
				&mediaItem.Latitude,
				&mediaItem.Longitude,
				&mediaItem.FPS,
				&mediaItem.EXIFData,
				&mediaItem.Keywords,
				&mediaItem.CreatedAt,
				&mediaItem.UpdatedAt); err != nil {
				slog.Error("error scanning mediaitem", "error", err)
				return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
			}
			mediaItems = append(mediaItems, mediaItem)
		}
		return ctx.JSON(http.StatusOK, mediaItems)
	}
	rows, err := h.DB.Query(ctx.Request().Context(), "SELECT * FROM mediaitems WHERE to_tsvector('english', keywords) @@ plainto_tsquery('english', $1) LIMIT $2", searchQuery, searchDefaultLimit)
	if err != nil {
		slog.Error("error searching mediaitems", "error", err)
		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
	}
	defer rows.Close()
	for rows.Next() {
		mediaItem := models.MediaItem{}
		if err := rows.Scan(&mediaItem.ID,
			&mediaItem.UserID,
			&mediaItem.Filename,
			&mediaItem.Hash,
			&mediaItem.Description,
			&mediaItem.MimeType,
			&mediaItem.SourceURL,
			&mediaItem.PreviewURL,
			&mediaItem.ThumbnailURL,
			&mediaItem.Placeholder,
			&mediaItem.IsFavourite,
			&mediaItem.IsHidden,
			&mediaItem.IsDeleted,
			&mediaItem.Status,
			&mediaItem.MediaItemType,
			&mediaItem.MediaItemCategory,
			&mediaItem.Width,
			&mediaItem.Height,
			&mediaItem.CreationTime,
			&mediaItem.CameraMake,
			&mediaItem.CameraModel,
			&mediaItem.FocalLength,
			&mediaItem.ApertureFnumber,
			&mediaItem.IsoEquivalent,
			&mediaItem.ExposureTime,
			&mediaItem.Latitude,
			&mediaItem.Longitude,
			&mediaItem.FPS,
			&mediaItem.EXIFData,
			&mediaItem.Keywords,
			&mediaItem.CreatedAt,
			&mediaItem.UpdatedAt); err != nil {
			slog.Error("error scanning mediaitem", "error", err)
			return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
		}
		mediaItems = append(mediaItems, mediaItem)
	}
	return ctx.JSON(http.StatusOK, mediaItems)
}