# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smriti API
  description: Smarter Home for all your Photos and Videos
  version: 24.06.24
servers:
  - url: https://localhost:5001
...
paths:
...
  /v1/features:
    get:
      summary: Get list of enabled features
      description: Get list of enabled features
      operationId: getFeatures
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Features'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    Features:
      type: object
      properties:
        favourites:
          type: boolean
          example: true
        hidden:
          type: boolean
          example: false
...